nbft: make lookup_ctrl function public
To prepare for the addition of nbft functionality fixup the fabrics
declarations.
